v3.0: 161题全量AC验证 + 课程系统 + 课本代码同步
## 课程管理模块
- 新增课程页面 (CourseList/CourseDetail)，复用Contest基础设施
- 课程→章节→例题/习题三级结构，支持可见性控制
- 后台管理：课程/章节/题目CRUD + 移动/排序
- JSON持久化存储，不修改数据库结构

## 全量AC验证 (Python + C++ 双语言)
- 161/161 Python AC ✅
- 161/161 C++ AC ✅
- 测试数据完整性校验通过
- Bank ↔ OJ 测试数据同步

## 课本代码更新
- lessons_v2/lesson01-16/codes/ 全部321文件同步
- acwing_codes + algorithm_basic_codes 更新
- 16章HTML + combined_print + PDF + DOCX 重建

## 测试工具链
- scripts/test_chapter.py — 标准化章节测试
- scripts/verify_all.py — 三重验证(DB+测试数据+本地运行)
- scripts/init_courses.py — 课程初始化

## 其他修复
- 移除Contest分组功能 (恢复原始列表视图)
- 章节测试数据修复(Ch1-16)
- 代码输出格式修复

+++ acwing_codes/第一讲_变量、输入输出、表达式与顺序语句/习题/AcWing 612. 球的体积.cpp
@@ -1,11 +1,2 @@
 #include <cstdio>
-
-int main()
-{
-    double r;
-    scanf("%lf", &r);
-
-    printf("VOLUME = %.3lf\n", 4.0 / 3.0 * 3.14159 * r * r * r);
-
-    return 0;
-}
+int main() { int r; scanf("%d",&r); printf("VOLUME = %.3lf\n",4.0/3.0*3.14159*r*r*r); return 0; }
